project layout corrected
