--- src/main/java/appfunctions/carsmanagement/AbstractCar.java
@@ -1,33 +1,0 @@
-package appfunctions.carsmanagement;
-
-public abstract class AbstractCar {
-    protected String id;
-    protected String brand;
-    protected String model;
-
-    abstract boolean isNull();
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getBrand() {
-        return brand;
-    }
-
-    public void setBrand(String brand) {
-        this.brand = brand;
-    }
-
-    public String getModel() {
-        return model;
-    }
-
-    public void setModel(String model) {
-        this.model = model;
-    }
-}
