--- src/main/java/appfunctions/carsmanagement/Car.java
@@ -1,9 +1,0 @@
-package appfunctions.carsmanagement;
-
-public class Car extends AbstractCar {
-
-    @Override
-    boolean isNull() {
-        return false;
-    }
-}
