--- src/main/java/appfunctions/carsmanagement/DeleteCar.java
@@ -1,6 +1,0 @@
-package appfunctions.carsmanagement;
-
-public class DeleteCar {
-
-    //TUTAJ ZROBIE USUWANIE SAMOCHODU
-}
