--- src/main/java/appfunctions/carsmanagement/NullCar.java
@@ -1,8 +1,0 @@
-package appfunctions.carsmanagement;
-
-public class NullCar extends AbstractCar {
-    @Override
-    boolean isNull() {
-        return true;
-    }
-}
