--- src/main/java/appnavigation/menuchoice/ShowMenu.java
@@ -1,11 +1,0 @@
-package appnavigation.menuchoice;
-
-public class ShowMenu {
-    public static void showMenu() {
-        System.out.println("Text tytułowy");
-        System.out.println("1 - Moje auta");
-        System.out.println("2 - Druga opcja");
-        System.out.println("3 - Trzecia opcja");
-        System.out.println("q - Zakończ program");
-    }
-}
