+++ src/main/java/com/lumato/appfunctions/carsmanagement/Car.java
@@ -1,0 +1,9 @@
+package com.lumato.appfunctions.carsmanagement;
+
+public class Car extends AbstractCar {
+
+    @Override
+    boolean isNull() {
+        return false;
+    }
+}
