+++ src/main/java/com/lumato/appfunctions/carsmanagement/DeleteCar.java
@@ -1,0 +1,6 @@
+package com.lumato.appfunctions.carsmanagement;
+
+public class DeleteCar {
+
+    //TUTAJ ZROBIE USUWANIE SAMOCHODU
+}
