+++ src/main/java/com/lumato/appfunctions/carsmanagement/NullCar.java
@@ -1,0 +1,8 @@
+package com.lumato.appfunctions.carsmanagement;
+
+public class NullCar extends AbstractCar {
+    @Override
+    boolean isNull() {
+        return true;
+    }
+}
