+++ src/main/java/com/lumato/tools/CLS.java
@@ -1,0 +1,13 @@
+package com.lumato.tools;
+
+public class CLS {
+
+    public static void clearScreen(){
+
+        //wstępna, "upośledzona" wersja czyszczenia konsoli
+        for(int i=0; i<50; i++){
+            System.out.print("\n");
+        }
+    }
+
+}
