--- src/main/java/tools/CLS.java
@@ -1,13 +1,0 @@
-package tools;
-
-public class CLS {
-
-    public static void clearScreen(){
-
-        //wstępna, "upośledzona" wersja czyszczenia konsoli
-        for(int i=0; i<50; i++){
-            System.out.print("\n");
-        }
-    }
-
-}
