+++ src/main/resources/currentUser.txt
@@ -1,0 +1,1 @@
+kasia;2
