+++ src/main/resources/profiles/kasia/kasia_cars.txt
@@ -1,2 +1,2 @@
-1;Lokomotywa;Nadświetlna
-2;Kulig;DziadkaMroza
+1;Lokomotywa;Parowa
+2;aaa;bbb
