+++ src/main/resources/profiles/marek/marek_cars.txt
@@ -1,3 +1,3 @@
 1;Chonda;Siwik
-2;Opel;Astra
-3;Drezyna;Reczna
+2;Drezyna;naWode
+1;Opel;Agila
