--- src/main/resources/test.txt
@@ -1,1 +1,0 @@
-test word
