--- src/main/test.csv
@@ -1,13 +1,0 @@
-AA;BB;CC
-1;10;100
-2;20;200
-3;30;300
-4;40;400
-5;50;500
-6;60;600
-7;70;700
-8;80;800
-text1;text2;text3;;text4text10;
-a1;a2;a3;a4
-b1;b2;b3;b4
-c1;
