--- src/main/test.txt
@@ -1,14 +1,0 @@
-Siała baba mak
-Nie wiedziała jak
-asadaasda
-Ale to juz bylo
-Wiersz testowy nr 4
-Ale to juz było
-MOze to jest to
-aaa
-alal
-aca
-ale to kuz; akkkajj 
-Zobaczymy co to bedzie teraz ;
-a teraz?
-user input; is nice
